chore(Linting): fix linting issues
