+++ gamma_cloudinary/templatetags/gamma_cloudinary_static.py
@@ -1,8 +1,8 @@
+from cloudinary import CloudinaryResource
 from django import template
 from django.utils.safestring import mark_safe
 from django.contrib.staticfiles.storage import staticfiles_storage
-
-from cloudinary import CloudinaryResource
+from gamma_cloudinary.storage.helpers import get_cloudinary_resource_type
 
 register = template.Library()
 
@@ -18,6 +18,6 @@ def gamma_cloudinary_static(context, resource_name, options_dict={}, **options):
         resource_path = staticfiles_storage.upload_path(resource_name)
         resource = CloudinaryResource(
             resource_path,
-            default_resource_type=staticfiles_storage._get_resource_type(resource_path)
+            default_resource_type=get_cloudinary_resource_type(resource_path)
             )
     return mark_safe(resource.build_url(**options))
