+++ tests/settings.py
@@ -46,8 +46,8 @@ STATICFILES_DIRS = [
 ]
 
 CLOUDINARY_STORAGE = {
-    'CLOUD_NAME': 'test', 
+    'CLOUD_NAME': 'test',
     'API_KEY': 'test_api_key',
     'API_SECRET': 'test_api_secret',
-    'SECURE': True  
+    'SECURE': True
 }
