+++ tests/static/css/bar.css
@@ -1,3 +1,3 @@
-*{
+* {
     color: red;
 }
