+++ tests/static/css/foo.css
