+++ tests/static/css/unmodified.css
@@ -1,3 +1,3 @@
-*{
+* {
     color: green;
 }
