+++ tests/test_cloudinary_storage.py
@@ -1,7 +1,6 @@
 from unittest.mock import patch
 from requests.exceptions import HTTPError
 from django.test import SimpleTestCase
-from django.core.validators import URLValidator
 from django.core.files.base import ContentFile
 from gamma_cloudinary.storage.CloudinaryStorage import CloudinaryStorage
 from .helpers import mock_http_response
@@ -9,7 +8,7 @@ from .helpers import mock_http_response
 class CloudinaryStorageTestCase(SimpleTestCase):
     def setUp(self):
         self.storage = CloudinaryStorage()
-    
+
     def test_upload_path(self):
         filename = 'css/test.css'
         prefixed_name = self.storage.upload_path(filename)
@@ -53,7 +52,7 @@ class CloudinaryStorageTestCase(SimpleTestCase):
         file_name = 'css/test.css'
         mock_url.return_value = 'https://res.cloudinary.com/cloudname/raw/upload/v1/files/css/test.css'
         mock_http_get.return_value= mock_http_response(
-            status=500, 
+            status=500,
             raise_for_status=HTTPError("Http error")
             )
         self.assertRaises(HTTPError, self.storage._open, file_name)
@@ -64,7 +63,7 @@ class CloudinaryStorageTestCase(SimpleTestCase):
         file_name = 'css/test.css'
         mock_url.return_value = 'https://res.cloudinary.com/cloudname/raw/upload/v1/files/css/test.css'
         mock_http_head.return_value= mock_http_response(
-            status=404, 
+            status=404,
             raise_for_status=HTTPError("Http error")
             )
         self.assertFalse(self.storage.exists(file_name))
@@ -75,7 +74,7 @@ class CloudinaryStorageTestCase(SimpleTestCase):
         file_name = 'css/test.css'
         mock_url.return_value = 'https://res.cloudinary.com/cloudname/raw/upload/v1/files/css/test.css'
         mock_http_head.return_value= mock_http_response(
-            status=500, 
+            status=500,
             raise_for_status=HTTPError("Http error")
             )
         self.assertRaises(HTTPError, self.storage.exists, file_name)
@@ -86,16 +85,9 @@ class CloudinaryStorageTestCase(SimpleTestCase):
         file_name = 'css/test.css'
         mock_url.return_value = 'https://res.cloudinary.com/cloudname/raw/upload/v1/files/css/test.css'
         mock_http_head.return_value= mock_http_response(
-            status=200, 
+            status=200,
             )
         self.assertTrue(self.storage.exists(file_name))
-
-    def test__get_resource_type_returns_the_right_resource_type(self):
-        self.assertEqual(self.storage._get_resource_type('css/test.css'), 'raw')
-        self.assertEqual(self.storage._get_resource_type('images/image.jpg'), 'image')
-        self.assertEqual(self.storage._get_resource_type('images/test.svg'), 'raw')
-        self.assertEqual(self.storage._get_resource_type('videos/test.mp4'), 'video')
-        self.assertEqual(self.storage._get_resource_type('audio/test.mp3'), 'video')
